+++ src/com/tutorial/game/Trail.java
@@ -30,7 +30,7 @@ public class Trail extends GameObject {
         g2d.setComposite(makeTransparent(alpha));
 
         g.setColor(color);
-        g.fillRect(x, y, width, height);
+        g.fillRect((int) x, (int) y, width, height);
 
         g2d.setComposite(makeTransparent(1));
     }
